+++ admin/src/components/CMEditView/RightLinksCompo/Summary/PreviewChecks/index.js
@@ -33,7 +33,7 @@ const SeoChecker = ({ checks }) => {
       </Typography>
 
       <Box paddingTop={4}>
-        <Stack size={2} key="good" horizontal background="neutral0">
+        <Stack spacing={2} key="good" horizontal background="neutral0">
           <Icon
             aria-hidden={true}
             colors={(theme) => ({
@@ -51,7 +51,7 @@ const SeoChecker = ({ checks }) => {
             : {good}
           </Typography>
         </Stack>
-        <Stack size={2} key="improvements" horizontal background="neutral0">
+        <Stack spacing={2} key="improvements" horizontal background="neutral0">
           <Icon
             aria-hidden={true}
             colors={(theme) => ({
@@ -69,7 +69,7 @@ const SeoChecker = ({ checks }) => {
             : {improvements}
           </Typography>
         </Stack>
-        <Stack size={2} key="bad" horizontal background="neutral0">
+        <Stack spacing={2} key="bad" horizontal background="neutral0">
           <Icon
             aria-hidden={true}
             colors={(theme) => ({
