+++ admin/src/components/CMEditView/RightLinksCompo/Summary/reducer.js
@@ -6,8 +6,8 @@ function reducer(state, action) {
       return {
         ...state,
         [action.value.entity]: {
-          color: action.value.color,
-          message: action.value.message,
+          color: action.value?.color,
+          message: action.value?.message,
         },
       };
     case 'UPDATE_FOR_PREVIEW':
