+++ admin/src/translations/en.json
@@ -14,20 +14,34 @@
   "SEOPage.info.create-single-type": "Create your first single-type",
   "SEOPage.info.added": "Added",
   "SEOPage.info.add": "Add component",
+  "SEOPage.info.settings": "Settings",
+  "SEOPage.info.settings.meta-title-check": "Meta Title",
+  "SEOPage.info.settings.meta-description-check": "Meta Description",
+  "SEOPage.info.settings.meta-robots-check": "Meta Robots",
+  "SEOPage.info.settings.meta-social-check": "Meta Social",
+  "SEOPage.info.settings.word-count-check": "Word Count",
+  "SEOPage.info.settings.canonical-url-check": "Canonical URL",
+  "SEOPage.info.settings.keyword-density-check": "Keyword Density",
+  "SEOPage.info.settings.structured-data-check": "Structured Data",
+  "SEOPage.info.settings.alternative-text-check": "Alt Text",
+  "SEOPage.info.settings.last-updated-at-check": "Last Updated At",
+  "SEOPage.info.settings.cancel.button": "Cancel",
+  "SEOPage.info.settings.save.button": "Save",
+  "SEOPage.info.settings.information": "Disable SEO checks for this specific content-type.",
   "SEOPage.tab.collection-type-title": "Collection Types",
   "SEOPage.tab.single-type-title": "Single Types",
   "SEOPage.tab-panel.column-name": "Name",
   "SEORightLinks.summary-title": "SEO Summary",
   "SEOSocialPreview.website-url": "url-of-your-website.io",
-  "SEOModal.summary-title.meta-title": "Meta title",
-  "SEOModal.summary-title.meta-description": "Meta description",
-  "SEOModal.summary-title.word-counter": "Word counter",
+  "SEOModal.summary-title.meta-title": "Meta Title",
+  "SEOModal.summary-title.meta-description": "Meta Description",
+  "SEOModal.summary-title.word-counter": "Word Counter",
   "SEOModal.summary-title.word-densisty": "Keyword Density",
-  "SEOModal.summary-title.meta-social": "Meta Social tag",
+  "SEOModal.summary-title.meta-social": "Meta Social Tags",
   "SEOModal.summary-title.canonical-url": "Canonical URL",
   "SEOModal.summary-title.structured-data": "JSON Structured Data",
   "SEOModal.summary-title.meta-robots": "Meta Robots",
-  "SEOModal.summary-title.alt-attributes": "Alt",
+  "SEOModal.summary-title.alt-attributes": "Alternative Texts",
   "SEOModal.summary-title.last-updated-at": "Last updated at",
   "SEOChecks.keywordsDensityCheck.default": "Every keyword are used more than 2 times.",
   "SEOChecks.keywordsDensityCheck.no-keywords": "No keywords were found in your SEO component.",
@@ -59,7 +73,7 @@
   "SEOChecks.metaSocialCheck.label": "Meta social tags allow you to manage the title, description & image of your posts.",
   "SEOChecks.metaTitleCheck.default": "A Meta Title has been found!",
   "SEOChecks.metaTitleCheck.not-found": "No Meta Title has been found.",
-  "SEOChecks.structuredDataCheck.default": "A Structured Data JSON has been found! However, we can validate the accuracy of its content.",
+  "SEOChecks.structuredDataCheck.default": "A Structured Data JSON has been found! However, we can't validate the accuracy of its content.",
   "SEOChecks.structuredDataCheck.not-found": "No Structured Data JSON has been found.",
   "SEOChecks.structuredDataCheck.label": "Structured data is a standardized format for providing information about a page and classifying the page content.",
   "SEOChecks.wordCountCheck.default": "You have more than 300 words. The word count is not a direct ranking factor. But your content should be as high quality as possible, with relevant and unique information. To meet these conditions, your content requires a minimum of paragraphs, and therefore of words.",
@@ -90,6 +104,6 @@
   "Good": "Good",
   "Improvements": "Improvements",
   "Bad": "Bad",
-  "Social-preview.alert": "Complete you social component to see the preview",
+  "Social-preview.alert": "Complete your social component to see the preview",
   "Information": "Information"
 }
