+++ admin/src/translations/fr.json
@@ -14,21 +14,35 @@
   "SEOPage.info.create-single-type": "Créer votre premier single-type",
   "SEOPage.info.added": "Ajouté",
   "SEOPage.info.add": "Ajouter le composant",
+  "SEOPage.info.settings": "Paramètres",
+  "SEOPage.info.settings.meta-title-check": "Meta Titre",
+  "SEOPage.info.settings.meta-description-check": "Meta Description",
+  "SEOPage.info.settings.meta-robots-check": "Meta Robots",
+  "SEOPage.info.settings.meta-social-check": "Meta Social",
+  "SEOPage.info.settings.word-count-check": "Compteur de mots",
+  "SEOPage.info.settings.canonical-url-check": "URL Canonique",
+  "SEOPage.info.settings.keyword-density-check": "Densité de mots-clés",
+  "SEOPage.info.settings.structured-data-check": "Données Structurées",
+  "SEOPage.info.settings.alternative-text-check": "Alt text",
+  "SEOPage.info.settings.last-updated-at-check": "Date de dernière mise à jour",
+  "SEOPage.info.settings.cancel.button": "Annuler",
+  "SEOPage.info.settings.save.button": "Enregistrer",
+  "SEOPage.info.settings.information": "Désactiver des verifications SEO pour ce content-type.",
   "SEOPage.tab.collection-type-title": "Types de collections",
   "SEOPage.tab.single-type-title": "Types simples",
   "SEOPage.tab-panel.column-name": "Nom",
   "SEORightLinks.summary-title": "Résumé du SEO",
   "SEOSocialPreview.website-url": "url-of-your-website.io",
-  "SEOModal.summary-title.meta-title": "Titre méta",
-  "SEOModal.summary-title.meta-description": "Meta description",
-  "SEOModal.summary-title.word-counter": "Word counter",
+  "SEOModal.summary-title.meta-title": "Meta Titre",
+  "SEOModal.summary-title.meta-description": "Meta Description",
+  "SEOModal.summary-title.word-counter": "Word Counter",
   "SEOModal.summary-title.word-densisty": "Densité des mots-clés",
   "SEOModal.summary-title.meta-social": "Balise Meta Social",
   "SEOModal.summary-title.canonical-url": "URL canonique",
-  "SEOModal.summary-title.structured-data": "Données structurées JSON",
+  "SEOModal.summary-title.structured-data": "Données Structurées JSON",
   "SEOModal.summary-title.meta-robots": "Meta Robots",
-  "SEOModal.summary-title.alt-attributes": "Alt",
-  "SEOModal.summary-title.last-updated-at": "Dernière mise à jour à",
+  "SEOModal.summary-title.alt-attributes": "Texte d'images",
+  "SEOModal.summary-title.last-updated-at": "Dernière mise à jour",
   "SEOChecks.keywordsDensityCheck.default": "Chaque mot-clé est utilisé plus de 2 fois.",
   "SEOChecks.keywordsDensityCheck.no-keywords": "Aucun mot-clé n'a été trouvé dans votre composant SEO.",
   "SEOChecks.keywordsDensityCheck.one-not-used": "Un mot-clé n'est pas utilisé dans votre contenu.",
