+++ server/services/index.js
@@ -1,7 +1,9 @@
 'use strict';
 
 const seo = require('./seo');
+const settings = require('./settings');
 
 module.exports = {
   seo,
+  settings
 };
